refactor: 도메인 상수를 도메인별 파일로 분리 #88
291줄에 회의·액션·인수인계·기업 상수가 다 있었다. domain.ts는 모아 내보내는
자리로 남겨 기존 import는 그대로 둔다.
